requirements.txt
requests>=2.28.1
Pillow>=9.2.0
torch>=1.13.1
numpy>=1.23.4
fal_client 